[torch.compile] Allow usage of Opaque Objects in PyTorch 2.11
We turned this off temporarily in the pt2.11 upgrade because there was
some failing tests and I wanted the pt2.11 upgrade to go in first.
This PR turns Opaque Objects back on and fixes the tests.

The problem with the tests was that we need to monkeypatch inductor to
fix an issue with opaque objects. We were only monkeypatching inductor
in the VLLM_COMPILE path (not in the DYNAMO_ONCE/STOCK_TORCH_COMPILE)
paths. This PR makes it so that the patch is global.

Test Plan:
- wait for tests
